Revert dark theme values for bg_1 and bg_2
As of this commit, the values in the selenized git repository for the
bg_1 and bg_2 colors are incorrect. An issue describing the problem,
along with a pull request to fix it, were opened on 2022-01-31 but
remain open as of this commit (see
https://github.com/jan-warchol/selenized/issues/99 and
https://github.com/jan-warchol/selenized/pull/100).

The upstream repository this repository is forked from contained the
corrected values for bg_1 and bg_2. However, commit 5940ba4 (Fix cursor
and tab colors, 2024-11-01) reverted the values for bg_1 and bg_2 in the
dark theme back to original broken values. This commit reverts those
changes.